Update Dashboard.Rmd
I added explanations to both the Core Findings page and the Bib. Archive page because A) I wasn't sure it would be obvious that you can interact with the table, and B) I wanted to have an explanation for the SoE scores somewhere on the page itself. I also made a few other minor aesthetic changes (like removing the length argument from the Datatables to save some space, etc).